requirements.txt
fastapi>=0.115,<1
uvicorn>=0.30,<1
pydantic>=2.7,<3
sqlalchemy>=2.0,<3
psycopg[binary]>=3.2,<4
google-genai>=1,<2
groq>=0.18,<1
Pillow>=10,<12
